comment page set up

+++ src/main/java/io/braxton/topicvpterjpa/interfaces/TopicRepoistory.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TopicRepoistory extends CrudRepository<Topic, Integer> {
+public interface TopicRepoistory extends CrudRepository<Topic, Long> {
 
 }
